+++ Website/src/app/components/base_layout.tsx
@@ -1,6 +1,6 @@
 import "./globals.css";
-import Header from "@/app/components/header";
-import Footer from "@/app/components/footer";
+import Header from "@/app/components/layout/header";
+import Footer from "@/app/components/layout/footer";
 import React from "react";
 
 /**
